Rename InfoPanels to HoverPanels.
