+++ mettascope/index.css
@@ -1,13 +1,17 @@
 /* Generated CSS from Figma file: https://www.figma.com/design/WeQldl3PsqFIpDnTka5Kd3 by tools/gen_html.py */
 
 
-*, *::before, *::after {
+*,
+*::before,
+*::after {
     margin: 0;
     padding: 0;
     border: none;
     box-sizing: border-box;
 }
-html, body {
+
+html,
+body {
     font-family: sans-serif;
 }
 
@@ -1769,7 +1773,7 @@ button#help-button.hover-icon .icon {
     overflow: hidden;
 }
 
-.infopanel {
+.hover-panel {
     position: absolute;
     display: flex;
     height: fit-content;
@@ -1785,7 +1789,7 @@ button#help-button.hover-icon .icon {
     overflow: hidden;
 }
 
-.infopanel .panel-body {
+.hover-panel .panel-body {
     position: relative;
     display: flex;
     height: fit-content;
@@ -2007,7 +2011,7 @@ button.memory .text {
     line-height: 26px;
 }
 
-.infopanel .tip {
+.hover-panel .tip {
     position: relative;
     display: block;
     width: 32px;
@@ -2015,7 +2019,7 @@ button.memory .text {
     background-color: rgba(0, 0, 0, 0.00);
 }
 
-.infopanel .close.hover-icon {
+.hover-panel .close.hover-icon {
     position: absolute;
     display: block;
     width: 32px;
