+++ mettascope/index.html
@@ -197,7 +197,7 @@
         <img src="data/ui/hide.png" alt="icon" class="icon">
       </div>
     </div>
-    <div class="infopanel">
+    <div class="hover-panel">
       <div class="panel-body">
         <div class="params">
           <img src="data/ui/tack.png" alt="pin-info-panel.hover-icon" class="pin-info-panel hover-icon">
